Moves auxiliary helper methods from `gmm_inference.py` to their own module in `inference/gmm_inference_utils.py` to clean up the `gmm_inference.py` module. Additionally, `inference.py` is renamed to `gauge_inference.py`.
